Fix Edit and Save
