--- extensions/blocks/story/story.js
@@ -1,42 +1,0 @@
-/**
- * External dependencies
- */
-import classNames from 'classnames';
-import { useEffect, useRef } from '@wordpress/element';
-
-/**
- * Internal dependencies
- */
-import player from './player';
-
-const storyPlayerSettings = {
-	slides: [],
-	shadowDOM: { enabled: false },
-	playInFullscreen: false,
-	tapToPlayPause: true,
-};
-
-export default function Story( { className, isSelected, mediaFiles, mountPlayer = true } ) {
-	let mountRef;
-	// cannot call useRef for save()
-	if ( mountPlayer ) {
-		mountRef = useRef();
-
-		useEffect( () => {
-			// render player asynchronously to avoid interacting with its UI
-			// when focusing on the block
-			setTimeout( () => {
-				if ( ! mountRef.current ) {
-					return;
-				}
-				player( mountRef.current, {
-					...storyPlayerSettings,
-					slides: mediaFiles,
-					disabled: ! isSelected,
-				} );
-			}, 0 );
-		}, [ mediaFiles, isSelected ] );
-	}
-
-	return <div className={ classNames( `wp-story`, className ) } ref={ mountRef }></div>;
-}
